@@ -25,28 +25,31 @@
     "deploy": "firebase deploy --project ma-cave-15a66"
   },
   "dependencies": {
-    "@angular/cli": "6.1.0-rc.3",
-    "@angular/common": "6.1.0-rc.3",
-    "@angular/compiler": "6.1.0-rc.3",
-    "@angular/compiler-cli": "6.1.0-rc.3",
-    "@angular/core": "6.1.0-rc.3",
-    "@angular/forms": "6.1.0-rc.3",
-    "@angular/http": "6.1.0-rc.3",
-    "@angular/platform-browser": "6.1.0-rc.3",
-    "@angular/platform-browser-dynamic": "6.1.0-rc.3",
-    "@angular/platform-server": "6.1.0-rc.3",
+    "@angular/cli": "6.1.0",
+    "@angular/common": "6.1.0",
+    "@angular/compiler": "6.1.0",
+    "@angular/compiler-cli": "6.1.0",
+    "@angular/core": "6.1.0",
+    "@angular/forms": "6.1.0",
+    "@angular/http": "6.1.0",
+    "@angular/platform-browser": "6.1.0",
+    "@angular/platform-browser-dynamic": "6.1.0",
+    "@angular/platform-server": "6.1.0",
     "@angular/tsc-wrapped": "4.4.7",
+    "@ionic/angular": "4.0.0-beta.0",
     "@ionic/storage": "2.1.3",
+    "@ionic/v4-migration-tslint": "1.4.0",
     "@ngrx/effects": "6.0.1",
     "@ngrx/store": "6.0.1",
     "@ngrx/store-devtools": "6.0.1",
     "@ngx-translate/core": "10.0.2",
     "@ngx-translate/http-loader": "3.0.1",
     "angularfire2": "5.0.0-rc.11",
-    "bootstrap": "4.1.2",
+    "bootstrap": "4.1.3",
     "chart.js": "2.7.2",
-    "firebase": "5.2.0",
-    "firebaseui": "3.1.1",
+    "codelyzer": "4.4.2",
+    "firebase": "5.3.0",
+    "firebaseui": "3.2.0",
     "firebaseui-angular": "3.1.1",
     "ionic-angular": "3.9.2",
     "lodash": "4.17.10",
@@ -61,20 +64,19 @@
   },
   "devDependencies": {
     "@ionic/app-scripts": "3.1.11",
-    "@types/bluebird": "3.5.21",
+    "@types/bluebird": "3.5.23",
     "@types/jasmine": "2.8.8",
     "@types/karma": "1.7.5",
-    "@types/node": "10.5.2",
+    "@types/node": "10.5.3",
     "bread-compressor-cli": "1.0.4",
     "http-server": "0.11.1",
-    "ionic": "3.20.0",
     "jasmine-core": "3.1.0",
     "jasmine-spec-reporter": "4.2.1",
     "ngrx-store-freeze": "0.2.4",
     "promise-polyfill": "8.0.0",
     "purify-css": "1.2.5",
     "rev-hash": "2.0.0",
-    "typescript": ">=2.7.2 and <2.8.0",
+    "typescript": "^2.9.0",
     "webpack-manifest-plugin": "2.0.3",
     "workbox-cli": "3.4.1"
   },
