@@ -1,5 +1,5 @@
 import {CUSTOM_ELEMENTS_SCHEMA, NgModule} from '@angular/core';
-import {IonicModule, IonicPageModule} from 'ionic-angular';
+import {IonicModule} from '@ionic/angular';
 import {TranslateModule} from '@ngx-translate/core';
 import {PopoverPage} from './popover-page';
 import {CommonModule} from '@angular/common';
@@ -7,7 +7,6 @@ import {CommonModule} from '@angular/common';
 @NgModule({
             imports: [
               IonicModule,
-              IonicPageModule.forChild(PopoverPage),
               CommonModule,
               TranslateModule
             ],
