@@ -1,22 +1,25 @@
 import {Component} from '@angular/core';
-import {IonicPage, ViewController} from 'ionic-angular';
 import {SearchCriteria} from '../../../model/search-criteria';
 import {Observable} from 'rxjs';
+import {ViewController} from '@ionic/core';
 
-@IonicPage()
 @Component({
              template: `
                <div class="inner">
                  <ion-list>
                    <ion-list-header>
+                     <ion-label></ion-label>
                      <ion-title>{{'dashboard.usual-queries' | translate}}</ion-title>
                    </ion-list-header>
                    <div *ngFor="let item of items$ | async" style="display: flex; align-items: center;">
-                     <button ion-item (click)="choose(item)" class="flex-grow">
-                       <span class="flex-grow">{{present(item.keywords)}}</span>
-                     </button>
-                     <cav-icon name="trash" (click)="removeSearch($event, item)"></cav-icon>
-                     <!--<ion-icon name="trash" (click)="removeSearch($event, item)" style="padding:5px"></ion-icon>-->
+                     <ion-item>
+                       <ion-label></ion-label>
+                       <button (click)="choose(item)" class="flex-grow">
+                         <span class="flex-grow">{{present(item.keywords)}}</span>
+                       </button>
+                       <cav-icon name="trash" (click)="removeSearch($event, item)"></cav-icon>
+                       <!--<ion-icon name="trash" (click)="removeSearch($event, item)" style="padding:5px"></ion-icon>-->
+                     </ion-item>
                    </div>
                  </ion-list>
                </div>`
