@@ -1,10 +1,9 @@
 import {CUSTOM_ELEMENTS_SCHEMA, NgModule} from '@angular/core';
-import {IonicModule, IonicPageModule} from 'ionic-angular';
+import {IonicModule} from '@ionic/angular';
 import {CellarPage} from './cellar-page';
 import {SharedModule} from '../../components/shared.module';
 import {FormsModule} from '@angular/forms';
 import {TranslateModule} from '@ngx-translate/core';
-import {CreateLockerPage} from './create-locker/create-locker-page';
 import {CommonModule} from '@angular/common';
 
 @NgModule({
@@ -12,7 +11,6 @@ import {CommonModule} from '@angular/common';
               CommonModule,
               FormsModule,
               IonicModule,
-              IonicPageModule.forChild(CellarPage),
               SharedModule,
               TranslateModule.forChild()
             ],
